requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
pinecone>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
typing-extensions>=4.8.0

